+++ api/tests_v2/common_import.py
@@ -35,3 +35,20 @@ from common.paddle_api_benchmark import PaddleAPIBenchmarkBase
 from common.tensorflow_api_benchmark import TensorflowAPIBenchmarkBase
 from common.api_param import APIConfig
 from common.main import test_main, test_main_without_json
+
+
+def unsqueeze_short(short, long):
+    """
+    Unsqueeze the short shape to the same length of the long's.
+    For example: short is [16, 2048] and long is [16, 2048, 7, 7],
+    it will return [16, 2048, 1, 1].
+    """
+    short_extend = np.ones([len(long)], dtype=np.int32).tolist()
+    start = 0
+    for value in short:
+        for i in range(start, len(long)):
+            if long[i] == value:
+                short_extend[i] = value
+                start = i
+                break
+    return short_extend
