+++ api/tests_v2/elementwise.py
@@ -15,23 +15,6 @@
 from common_import import *
 
 
-def unsqueeze_short(short, long):
-    """
-    Unsqueeze the short shape to the same length of the long's.
-    For example: short is [16, 2048] and long is [16, 2048, 7, 7],
-    it will return [16, 2048, 1, 1].
-    """
-    short_extend = np.ones([len(long)], dtype=np.int32).tolist()
-    start = 0
-    for value in short:
-        for i in range(start, len(long)):
-            if long[i] == value:
-                short_extend[i] = value
-                start = i
-                break
-    return short_extend
-
-
 class ElementwiseConfig(APIConfig):
     def __init__(self):
         super(ElementwiseConfig, self).__init__('elementwise')
