+++ api/tests_v2/elementwise_with_axis.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from common_import import *
-from elementwise import unsqueeze_short
 
 
 class ElementwiseWithAxisConfig(APIConfig):
