+++ api/tests_v2/remainder.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from common_import import *
-from elementwise import unsqueeze_short
 
 
 class RemainderConfig(APIConfig):
